Rename styles.js to style-installer
