+++ src/style-installer.js
