+++ test/functional/test-runtime.js
@@ -26,7 +26,7 @@ import {
 import {platformFor} from '../../src/platform';
 import * as ext from '../../src/service/extensions-impl';
 import * as extel from '../../src/extended-element';
-import * as styles from '../../src/styles';
+import * as styles from '../../src/style-installer';
 import * as dom from '../../src/dom';
 import * as sinon from 'sinon';
